Create Swap two nibbles in a byte(31 MAY GFG-POTD)
Bitmasking & Shifting:

(n & 0x0F) extracts the right nibble (last 4 bits) of n. This is done using the bitwise AND (&) with 0x0F (which is 00001111 in binary).
(n & 0xF0) extracts the left nibble (first 4 bits) of n. This is done using the bitwise AND (&) with 0xF0 (which is 11110000 in binary).
Swapping:

(n & 0x0F) << 4 shifts the right nibble by 4 bits to the left, placing it in the position of the left nibble.
(n & 0xF0) >> 4 shifts the left nibble by 4 bits to the right, placing it in the position of the right nibble.
Combining Nibbles:

The left-shifted right nibble and right-shifted left nibble are combined using the bitwise OR operator (|) to form the final swapped number.


PLEASE ASSIGN ME THIS